refactor: move editors back to substation folder
