+++ src/editors/substation/substation-editor.ts
@@ -14,11 +14,16 @@ import { Menu } from '@material/mwc-menu';
 import { IconButton } from '@material/mwc-icon-button';
 import { ListItem } from '@material/mwc-list/mwc-list-item';
 
-import '../action-pane.js';
+import '../../action-pane.js';
 import './ied-editor.js';
 import './voltage-level-editor.js';
-import { newActionEvent, newWizardEvent, SCLTag, tags } from '../foundation.js';
-import { emptyWizard, wizards } from '../wizards/wizard-library.js';
+import {
+  newActionEvent,
+  newWizardEvent,
+  SCLTag,
+  tags,
+} from '../../foundation.js';
+import { emptyWizard, wizards } from '../../wizards/wizard-library.js';
 import {
   cloneSubstationElement,
   selectors,
