+++ src/editors/substation/voltage-level-editor.ts
@@ -14,7 +14,7 @@ import { Menu } from '@material/mwc-menu';
 import { IconButton } from '@material/mwc-icon-button';
 import { ListItem } from '@material/mwc-list/mwc-list-item';
 
-import '../action-pane.js';
+import '../../action-pane.js';
 import './bay-editor.js';
 import './ied-editor.js';
 import {
@@ -23,10 +23,15 @@ import {
   cloneSubstationElement,
   styles,
 } from './foundation.js';
-import { newActionEvent, newWizardEvent, SCLTag, tags } from '../foundation.js';
+import {
+  newActionEvent,
+  newWizardEvent,
+  SCLTag,
+  tags,
+} from '../../foundation.js';
 
 import { SubstationEditor } from './substation-editor.js';
-import { emptyWizard, wizards } from '../wizards/wizard-library.js';
+import { emptyWizard, wizards } from '../../wizards/wizard-library.js';
 
 function childTags(element: Element | null | undefined): SCLTag[] {
   if (!element) return [];
