+++ test/integration/zeroline/voltage-level-editor-wizarding-editing.test.ts
@@ -3,8 +3,8 @@ import { fixture, html, expect } from '@open-wc/testing';
 import '../../mock-wizard-editor.js';
 import { MockWizardEditor } from '../../mock-wizard-editor.js';
 
-import '../../../src/zeroline/voltage-level-editor.js';
-import { VoltageLevelEditor } from '../../../src/zeroline/voltage-level-editor.js';
+import '../../../src/editors/substation/voltage-level-editor.js';
+import { VoltageLevelEditor } from '../../../src/editors/substation/voltage-level-editor.js';
 import { WizardTextField } from '../../../src/wizard-textfield.js';
 
 describe('voltage-level-editor wizarding editing integration', () => {
