+++ test/unit/action-pane.test.ts
@@ -1,7 +1,7 @@
 import { expect, fixture, html } from '@open-wc/testing';
 
 import '../../src/action-pane.js';
-import '../../src/zeroline/bay-editor.js';
+import '../../src/editors/substation/bay-editor.js';
 import { ActionPane } from '../../src/action-pane.js';
 
 describe('action-pane', () => {
